refactor(#176): update PopupWindow migrations
