+++ src/pages/popup/layout/PopupWindow.utils/PopupWindowStorage.ts
@@ -2,7 +2,6 @@ import { TypeOf } from 'io-ts';
 import browser from 'webextension-polyfill';
 
 import { decodeStruct, type } from '../../../../lib/types';
-import { MigrationTask } from '../../../../migrations/migrations';
 
 const storageStruct = type.type({
 	/**
@@ -65,33 +64,3 @@ export class PopupWindowStorage {
 		return this.setData(mergedData);
 	};
 }
-
-// TODO: #181 use library for migrations
-export const PopupWindowStorageMigration: MigrationTask = {
-	version: 1,
-	async migrate(prevVersion) {
-		switch (prevVersion) {
-			case 0: {
-				// Migrate data from `localStorage`
-				const keyPrefix = 'PopupPage.tabSet#';
-				for (const key of Object.keys(localStorage)) {
-					// Skip not match keys
-					if (!key.startsWith(keyPrefix)) continue;
-
-					// Copy
-					const value = localStorage.getItem(key);
-					if (typeof value === 'string') {
-						// Cut prefix to get hash from key
-						const hash = key.slice(keyPrefix.length);
-
-						// Write data
-						await new PopupWindowStorage().setActiveTab(hash, value);
-					}
-
-					// Remove
-					localStorage.removeItem(key);
-				}
-			}
-		}
-	},
-};
